+++ app/src/main/java/com/marcinorlowski/fonty/demo/MainActivity.java
@@ -3,7 +3,7 @@ package com.marcinorlowski.fonty.demo;
 /*
  ******************************************************************************
  *
- * Copyright 2013-2017 Marcin Orłowski
+ * Copyright 2013-2017 Marcin Orlowski
  *
  * Licensed under the Apache License 2.0
  *