+++ CHANGES.md
@@ -1,6 +1,9 @@
 Changelog
 =========
 
+ * @dev
+   - `init()` is now `context()`
+
  * v1.2.0 (2017-04-23)
    - [#2] Added support for `TextInputLayout`