+++ CHANGES.md
@@ -1,7 +1,7 @@
 Changelog
 =========
 
- * @dev
+ * v2.0.0 (2017-04-24)
    - Renamed `init()` method to `context()`
    - Configuration chain now must end with `done()`
    - Improved public methods argument validation